Update database views to calculate the service area
I updated the Views that we have so they model the behavior for
- Elected Officials (EO) service area and for the
- Management and Finance (MF) burea, which actually falls under LA
although it's budgeted for many service areas

So, if you query the view, it matches the amounts shown in the Budget in
Brief
SELECT fiscal_year, service_area_code, sum
FROM public."HistorySummaryByServiceArea"
Where fiscal_year = '2014-15';

1) I added the views through the pgAdmin3 console.
I originally created the views so I could do some prototyping and to
better understand the data.
If we use these views for anything other than that it's probably be
better to push these views to the db through the Django migration.

2) It seems like PostGres added the "order by" to the views. We don't
need an "order by" in the views.
From what I read, PostGres sort of ignores the "order by" in the views.
But for the most part, "order by" in a view is not good code.

3) We _could_ write Django endpoints that call the views instead of the
budget history table.

4) Or, we can do it differently: create a calculated column in the
database that reflects the logic above.
You can see the logic in the views.